requirements.txt
Jinja2==3.1.4
pydantic==2.11.4
Redis==6.1.0
setuptools==75.4.0
fastapi==0.115.12
uvicorn==0.34.2
asyncpg==0.30.0
tortoise-orm==0.25.0
aerich==0.9.0
pylint==2.17.5
pylint-exit==1.2.0
playwright==1.52.0
starlette==0.46.2
pytz==2025.2
fastapi-guard==2.1.2
APScheduler==3.11.0
dependency-injector==4.47.1